Fixed minor problems with stream setup UI (cherry picked from commit 72843d4)

+++ source.py
@@ -556,17 +556,18 @@ class Database(object):
         c.close()
         return expired
 
-
     def setCustomStreamUrl(self, channel, stream_url):
-        self._invokeAndBlockForResult(self._setCustomStreamUrl, channel, stream_url)
+        if stream_url is not None:
+            self._invokeAndBlockForResult(self._setCustomStreamUrl, channel, stream_url)
         # no result, but block until operation is done
 
     def _setCustomStreamUrl(self, channel, stream_url):
-        c = self.conn.cursor()
-        c.execute("DELETE FROM custom_stream_url WHERE channel=?", [channel.id])
-        c.execute("INSERT INTO custom_stream_url(channel, stream_url) VALUES(?, ?)", [channel.id, stream_url.decode('utf-8', 'ignore')])
-        self.conn.commit()
-        c.close()
+        if stream_url is not None:
+            c = self.conn.cursor()
+            c.execute("DELETE FROM custom_stream_url WHERE channel=?", [channel.id])
+            c.execute("INSERT INTO custom_stream_url(channel, stream_url) VALUES(?, ?)", [channel.id, stream_url.decode('utf-8', 'ignore')])
+            self.conn.commit()
+            c.close()
 
     def getCustomStreamUrl(self, channel):
         return self._invokeAndBlockForResult(self._getCustomStreamUrl, channel)
